+++ src/autogain/gstpeautogain.hpp
@@ -32,7 +32,7 @@ struct _GstPeautogain {
   float loudness;   // estimated loudness
   float gain;       // correction gain
   float range;      // loudness range
-  bool detect_silence, reset;
+  bool detect_silence, reset, use_geometric_mean;
 
   /* < private > */